# pipelineService/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
...
        "/data-products/": {
            "get": {
                "description": "Returns a list of all the data products",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data-products"
                ],
                "summary": "Returns all the data products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DataProductListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates the data product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data-products"
                ],
                "summary": "Create Data Product",
                "parameters": [
                    {
                        "description": "Product Info",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DataProduct"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.DataProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/data-products/internal/": {
            "get": {
                "description": "returns the name of the data product that can be transformed",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data-products/internal"
                ],
                "summary": "Returns updated data product",
                "parameters": [
                    {
                        "description": "Product Info",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DataProduct"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProductAssetDetails"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/data-products/internal/transformations/assets/": {
            "post": {
                "description": "deletes the already present assets and creates the new entries",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data-products/internal"
                ],
                "summary": "syncs the transformed assets with the db",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProductAssets"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/data-products/transformations/{id}": {
            "get": {
                "description": "Return DBT transformation details for a given product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data-products"
                ],
                "summary": "Return DBT transformation details for a given product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Apply transformation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data-products"
                ],
                "summary": "Apply transformation",
                "parameters": [
                    {
                        "description": "Source Details",
                        "name": "configureSourceData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateSourceConnectorRequestAPI"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Destination ID",
                        "name": "destinationId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Source ID",
                        "name": "sourceId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CreatePipelineRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/data-products/transformations/{id}/": {
            "put": {
                "description": "updates the data product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data-products"
                ],
                "summary": "Returns updated data product",
                "parameters": [
                    {
                        "description": "Transformation Info",
                        "name": "transformation",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdatePipelineAirByteRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreatePipelineRequest"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/data-products/{id}/": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns a data product by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data-products"
                ],
                "summary": "Returns a data product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetDataProductViewResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "updates the data product",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data-products"
                ],
                "summary": "Returns updated data product",
                "parameters": [
                    {
                        "description": "Product Info",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DataProduct"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DataProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/data-products/{id}/add-pipeline/": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "A newly created binding is returned",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "data-products"
                ],
                "summary": "Returns product and pipeline ID",
                "parameters": [
                    {
                        "description": "pipelines",
                        "name": "pipelines",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InputProductsPipelines"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProductsPipelinesResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.DataProduct": {
            "type": "object",
            "properties": {
                "dataDomain": {
                    "type": "string",
                    "example": "data-product-1"
                },
                "dataProductGovernance": {
                    "type": "string",
                    "example": "[sales, marketing]"
                },
                "dataProductStatus": {
                    "type": "string",
                    "example": "draft"
                },
                "description": {
                    "type": "string",
                    "example": "This data product will help you in creating better visualizations"
                },
                "lastUpdated": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "example": "data-product-1"
                },
                "owner": {
                    "type": "integer",
                    "example": 1
                },
                "productID": {
                    "type": "string",
                    "example": "a152379e-01a1-11ec-82d6-a312edcd9c7b"
                },
                "workspaceId": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.DataProductListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GetAllDataProductsView"
                    }
                },
                "errors": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errors": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}